Revert SemisupervisedTrainingMixin
